pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["artur sharipov <sharipoffartur@yandex.ru>"]
readme = "README.md"
packages = [{include = "page_analyzer"}]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.2.3"
gunicorn = "^20.1.0"
python-dotenv = "^0.21.1"
flake8 = "^6.0.0"
validators = "^0.20.0"
requests = "^2.28.2"
beautifulsoup4 = "^4.11.2"
flask-wtf = "^1.1.1"
psycopg2-binary = "^2.9.5"
psycopg2 = "^2.9.9"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

